Refactor Home Assistant datetime test stubs
